+++ packages/sdk-ui-angular/src/lib/services/index.ts
@@ -2,3 +2,4 @@ export * from './query.service';
 export * from './sisense-context.service';
 export * from './theme.service';
 export * from './dashboard.service';
+export * from './widget.service';