+++ packages/sdk-ui/vitest.config.ts
@@ -35,7 +35,6 @@ const config: UserConfig = {
         'src/@types',
         'src/widgets/common/drilldown-breadcrumbs',
         'src/charts/indicator/chart/*', // legacy indicator's code, copy-pasted from PWC
-        'src/ai/translators/*', // POC implementation of Chat to Code; will add unit tests in separate MRs
       ],
     },
   },